refactor(console): connect an agent in a dialog, manage connections in settings
The connections page mixed two different things and then had to apologise for
it in prose: "it signs in with your CLI login, so it does not appear in the
list below". A sentence explaining why the page is confusing is a sign the
page is wrong, not that the sentence was missing.

They are separated now. Connecting is a one-time task, so it is a dialog —
numbered steps, one command each, the client picker at the top. Which remote
clients hold access is something you manage over time, so it stays in Settings
as a real list.

That also removes the double rule under Agents. The section was a header with
no body, so its own bottom border landed directly on the divider after it. It
has content now, and the header rule separates the header from that content
the way it does in every other section.

Removes the /dashboard/settings/connections route and the page behind it, and
drops mcpConnectionCount, which nothing read any more but still cost a count
query on every dashboard load.
