+++ packages/console/app/components/dashboard-types.ts
@@ -108,8 +108,6 @@ export type DashboardInitialData = {
   billingEnabled: boolean;
   analyticsEnabled: boolean;
   remoteMcpOAuthEnabled: boolean;
-  /** Authorized remote MCP connections for the signed-in user. */
-  mcpConnectionCount: number;
 };
 
 export type ApiError = { error?: string };
