--- packages/console/app/dashboard/settings/connections/page.tsx
@@ -1,8 +1,0 @@
-'use client';
-
-import { McpConnectionsDashboard } from '@/app/components/McpConnectionsDashboard';
-import { useDashboardData } from '@/app/dashboard/DashboardDataProvider';
-
-export default function McpConnectionsPage() {
-  return <McpConnectionsDashboard initialData={useDashboardData()} />;
-}
